requirements.txt
easydict==1.8
Flask==1.0.2
gunicorn==19.9.0
matplotlib==3.0.0
numpy==1.15.1
opencv-python==3.4.3.18
pandas==0.23.4
Pillow==5.2.0
pluggy==0.7.1
protobuf==3.12.2
pytools==2018.5.2
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
tensorboard==1.10.0
tensorboardX==2.0
torch==1.2.0
torchvision==0.4.0
tqdm==4.40.0

